Move recorder chunk utils to shared collection utils
